refactor(e2e): group split modules into folders for clearer naming
Move kube-client, api-helper, common, rhdh-deployment, and semantic selector
splits into directories with short filenames, matching the ui-helper pattern.
