+++ e2e-tests/playwright/support/selectors/semantic/table-helpers.ts
@@ -1,6 +1,6 @@
 import { Page, Locator } from "@playwright/test";
 
-import { semanticSelectorsAccessibility } from "./semantic-selectors-accessibility";
+import { semanticSelectorsAccessibility } from "./accessibility";
 
 export function findTableCell(page: Page, rowText: string | RegExp, cellIndex: number): Locator {
   const row = semanticSelectorsAccessibility.tableRow(page, rowText);
