+++ e2e-tests/playwright/utils/authentication-providers/rhdh-deployment/k8s.ts
@@ -6,7 +6,7 @@ import * as k8s from "@kubernetes/client-node";
 import { expect } from "@playwright/test";
 import * as yaml from "yaml";
 
-import { hasErrorResponse } from "../errors";
+import { hasErrorResponse } from "../../errors";
 import {
   BackstageCr,
   currentDirName,
@@ -15,8 +15,8 @@ import {
   isRecord,
   RHDHDeploymentState,
   rootDirName,
-} from "./rhdh-deployment-types";
-import { ensureBackstageCRIsAvailable, waitForDeploymentReady } from "./rhdh-deployment-wait";
+} from "./types";
+import { ensureBackstageCRIsAvailable, waitForDeploymentReady } from "./wait";
 
 export async function readYamlToJson(filePath: string): Promise<unknown> {
   const fileContent = await fs.readFile(filePath, "utf8");
