+++ e2e-tests/playwright/utils/authentication-providers/rhdh-deployment/types.ts
