+++ e2e-tests/playwright/utils/common/auth-popup.ts
