+++ e2e-tests/playwright/utils/kube-client/diagnostics/pods.ts
@@ -5,7 +5,7 @@ import {
   formatContainerStartedAt,
   getKubeApiErrorMessage,
   podNameOrUnknown,
-} from "./kube-client-helpers";
+} from "../helpers";
 
 function logWaitingContainerStatus(
   containerName: string,
